chore: rename RoutinViewModel to RoutingViewModel
