+++ PingMe/Screen/Login/LoginViewModel.swift
@@ -1,4 +1,3 @@
-import CoreFoundation
 import Foundation
 import Observation
